Moving model-store into “generic” sub-directory.
Also adding a README.md to explain what belongs in “generic”
