+++ src/alerts/access-expiration-alert/hooks.js
@@ -1,7 +1,7 @@
 /* eslint-disable import/prefer-default-export */
 import { useMemo } from 'react';
-import { useModel } from '../../model-store';
 import { useAlert } from '../../user-messages';
+import { useModel } from '../../generic/model-store';
 
 export function useAccessExpirationAlert(courseId) {
   const course = useModel('courses', courseId);
