+++ src/course-home/dates-tab/Day.jsx
@@ -4,7 +4,7 @@ import classNames from 'classnames';
 import { useSelector } from 'react-redux';
 import { FormattedDate, injectIntl, intlShape } from '@edx/frontend-platform/i18n';
 
-import { useModel } from '../../model-store';
+import { useModel } from '../../generic/model-store';
 
 import { getBadgeListAndColor } from './badgelist';
 import { isLearnerAssignment } from './utils';
