+++ src/generic/README.md
@@ -1,0 +1,5 @@
+## Generic module (src/generic)
+
+This module is a place for code that is not specific to the frontend-app-learning micro-frontend.  It is intended to be reusable and we expect we can extract it to a reusable code repository like Paragon or frontend-platform in the future.
+
+Take care when changing this code that it doesn't take on domain-specific logic.
