+++ src/generic/model-store/hooks.js
