+++ src/store.js
@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { reducer as coursewareReducer } from './data';
-import { reducer as modelsReducer } from './model-store';
 import { reducer as courseHomeReducer } from './course-home/data';
+import { reducer as modelsReducer } from './generic/model-store';
 
 const store = configureStore({
   reducer: {
